pyproject.toml
[project]
name = "support-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Gershon Mac",email = "gershon.o@mblhightech.net"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "langchain (>=0.3.20,<0.4.0)",
    "langgraph (>=0.3.8,<0.4.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-core (>=0.3.44,<0.4.0)",
    "langchain-pinecone (>=0.2.3,<0.3.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "langchain-anthropic (>=0.3.9,<0.4.0)",
    "langchain-groq (>=0.2.5,<0.3.0)",
    "langchain-google-genai (>=2.1.0,<3.0.0)",
    "tavily-python (>=0.5.1,<0.6.0)",
    "streamlit (>=1.43.2,<2.0.0)",
    "audio-recorder-streamlit (>=0.0.10,<0.0.11)",
    "elevenlabs (>=1.54.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false